backend/python-ai/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
dashscope>=1.20.0
python-multipart>=0.0.9
PyPDF2>=3.0.0
python-docx>=1.1.0
aiofiles>=24.1.0
httpx>=0.27.0
